+++ lib/process_bot/process/runner.rb
@@ -119,6 +119,8 @@ class ProcessBot::Process::Runner
   end
 
   def stop_related_processes
+    ensure_subprocess_pgid_for_stop!
+
     loop do
       processes = related_processes
 
@@ -134,6 +136,12 @@ class ProcessBot::Process::Runner
         sleep 0.5
       end
     end
+  end
+
+  def ensure_subprocess_pgid_for_stop!
+    return if subprocess_pgid
+
+    raise "Unable to stop related processes because subprocess PGID could not be resolved (subprocess PID: #{subprocess_pid.inspect})"
   end
 
   def find_sidekiq_pid